r3_75: merged recent changes from r3_74.
git-svn-id: file:///home/apenwarr/alumnit-svn/public/branches/r3_75@3238 6619c942-59fa-0310-bf43-a07aaa81d9ac

+++ include/strutils.h
@@ -161,7 +161,7 @@ WvString getdirname(WvStringParm fullname);
  * Given a number of blocks and a blocksize (default==1 byte), return a 
  * WvString containing a human-readable representation of blocks*blocksize.
  */
-WvString sizetoa(long blocks, int blocksize=1);
+WvString sizetoa(long long blocks, int blocksize=1);
 
 /**
  * Finds a string in an array and returns its index.
